Demonstrate that a duplicate connection raises a PeerAlreadyConnectedException.
